Moved scrollNumberPickerToValue to ViewActions
